fix dune: include Generic in flow_typing_type
With this change, `dune build src/typing/flow_typing.a` works!

This module has an import cycle with flow_typing_type, so we can't
make it a separate Dune library as we usually would for source files
in different directories.

Similarly, other files in this directory have import cycles with
flow_typing.

Happily we can use `include_subdirs` to get these files included in
those two libraries, just as if they were in src/typing/ directly.