fix dune: type-sig
This becomes needed for `dune build src/flow.exe` after 34465eb6f,
but most of the changes correspond to the earlier d07107708, which
introduced some modules that simply weren't needed for the flow.exe
build until then.

It seems like some things are just missing, so we add to the stubs.