fix dune: handle flowlib accurately -- now all tests pass!
This makes the handling of flowlib in the Dune build more parallel to
how it works in the Makefile.

In particular, we identify flowlib files using a glob with `*.js`, in
order to specifically exclude lib/APACHE_LICENSE (which otherwise causes
parse errors as it's not JS.)  This becomes necessary after ea9d746d0.

Dune doesn't seem to quite have a way to declare a dependency on all
those files directly.  There's `glob_files`, but it doesn't seem to
be accepted in `preprocessor_deps`, only in plain `deps`.  But we can
get the same effect by adding a stamp rule that just takes the hash
of all the desired files.

The dune language bump is needed for the `target` option.